+++ CHANGELOG.md
@@ -8,7 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - BREAKING: Only send a subset of the fields sufficient for most use-cases to OPA for performance reasons.
   The old behavior of sending all fields can be restored by setting `hadoop.security.authorization.opa.extended-requests` to `true` ([#XX]).
-- Bump `okio` to 1.17.6 and to 3.9.1 afterwards  to get rid of CVE-2023-3635 ([#46], [#XX]).
+- Bump `okio` to 1.17.6 and to 3.9.1 afterwards to get rid of CVE-2023-3635 ([#46], [#XX]).
+
+### Fixed
+
+- Set path to `/` when the operation `contentSummary` is called on `/`. Previously path was set to `null` ([#XX]).
 
 [#46]: https://github.com/stackabletech/hdfs-utils/pull/46